requirements.txt
pyyaml==4.2b4
tqdm==4.28.1
grpcio==1.16.0
grpcio-tools==1.14.1
# mask_rcnn
numpy==1.16.0
scipy==1.2.0
Pillow==5.4.1
cython==0.29
matplotlib==3.0.2
scikit-image==0.14.2
tensorflow-gpu==1.12.0
keras==2.2.4
opencv-python==3.4.3.18
h5py==2.9.0
imgaug==0.2.6
IPython[all]==7.1.1